Consolidate SSHException log calls into one
The `except` block used to produce one log statement for the general
information and then one additional log statement for every line in the
stack trace since it was imported as a list.

The consolidated log call plays nicely with error reporting tools, which
would otherwise show this one error as several separate issues despite
having the same root cause.